Isolate `LogType` module
