+++ src/io.rs
@@ -17,7 +17,7 @@ pub use input::Input;
 pub use metadata::DeviceMetadata;
 pub use sensors::*;
 
-use crate::storage::{Container};
+use crate::storage::Container;
 
 /// Defines sensor type. Used to classify data along with `IOData`.
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
@@ -71,5 +71,4 @@ pub struct IOData {
     pub data: f64,
 }
 
-pub type LogType = Container<IOEvent, DateTime<Utc>>;
 pub type InputType = Box<dyn Input>;
