+++ src/storage.rs
@@ -2,9 +2,11 @@ mod collection;
 mod container;
 mod grouping;
 mod persistent;
+mod logging;
 
 pub use collection::MappedCollection;
 pub use container::{Container, Containerized};
 pub use grouping::PollGroup;
 pub use persistent::Persistent;
+pub use logging::LogType;
 pub use crate::io::IdType;
